BM Versi 2 -> pos and supplier product

+++ app/Http/Controllers/Api/Supplier/SupplierController.php
@@ -5,8 +5,12 @@ namespace App\Http\Controllers\Api\Supplier;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Supplier\Supplier;
+use App\Models\Products\SupplierProduct;
+use App\Models\Purchases\Purchase;
 use JWTAuth;
 use App\Http\Resources\Supplier\SupplierList as SupplierResource;
+use App\Http\Resources\Supplier\SupplierItem as SupplierItem;
+use App\Http\Resources\Supplier\SupplierProductList as SupplierProductResoure;
 
 
 class SupplierController extends Controller
@@ -30,6 +34,18 @@ class SupplierController extends Controller
            ],200);
     }
 
+    public function detail($id)
+    {
+       $supplier = Supplier::find($id);
+       $supplier_product = SupplierProduct::where('supplier_id',$id)->get();
+       return response()->json([
+        'success' => true,
+        'suppliers' =>  new SupplierItem($supplier),
+        'supplier_product' => new SupplierProductResoure($supplier_product)
+       ],200);
+
+    }
+
     public function update(Request $request,$id)
     {
         $supplier = Supplier::find($id);
@@ -42,11 +58,21 @@ class SupplierController extends Controller
 
     public function delete(Request $request,$id)
     {
-        $supplier = Supplier::find($id);
-        $supplier->delete();
-        return response()->json([
-            'success' => true,
-            'message' =>  "Supplier Berhasil dihapus"
-           ],200);
+        $sp = SupplierProduct::where('supplier_id',$id)->get();
+        $pc = Purchase::where('supplier_id',$id)->get();
+        if(empty($sp) && empty($sp)){
+            $supplier = Supplier::find($id);
+            $supplier->delete();
+            return response()->json([
+                'success' => true,
+                'message' =>  "Supplier Berhasil dihapus"
+               ],200);
+        }else{
+            return response()->json([
+                'success' => false,
+                'message' =>  "Supplier Tidak dapat dihapus"
+               ],400);
+        }
+
     }
 }
